Remove the -outputstate command line argument.
It was now essentially useless insofar as the dual command -inputstate was
removed in 8.14. Both command had been deprecated at that time already, so
I think it was an oversight from the PR that performed the removal.

The only use case I can think of would have relied on the unmarshalling of
arbitrary OCaml data using the Coq process, but this seems very far-fetched.