Move joining of futures outside of the API.
The only client joining futures was now the opaque table mechanism. In one
of the two remaining calls, joining was actually useless because the future
was being created locally and thus unreachable from any other piece of code.

This is actually suspicious since that means in practice joining a environment
is not enough to guarantee marshalling is going to result in a meaningful value.
This should be investigated.