Use a uniform representation for the internal persistent arrays.
We provide a low-level API that statically ensures that the arrays we
manipulate never use the flat-float representation, regardless of the
OCaml compiler flags. Hopefully this is designed carefully enough that
this invariant is really respected so that the VM and native do not
break the OCaml runtime semantics.

Fixes #15070: Memory corruption when mixing primitive floats and arrays.

(cherry picked from commit 9eb4d8c4997241f91c731a3844b79c44096bcd44)