S2 in the conclusion of WF-MOD2 is not optional.
The conclusion of WF-MOD2 was WF(E; Mod(X:S1[:=S2]))[].

But S2 should not be optional.

In the case that S2 is not given here,
WF-MOD2 is same as WF-MOD1 with meaningless premises for S2.

https://github.com/coq/coq/commit/81f12192810bdf825cee82658a36214740d1a75b
changes the macro \Mod as:

-\newcommand{\Mod}[3]{{\sf Mod}({#1}:{#2}:={#3})}
+\newcommand{\Mod}[3]{{\sf Mod}({#1}:{#2}\,\zeroone{:={#3}})}

So, "[" and "]" added in WF-MOD2 seems accidental.

(cherry picked from commit d5b254cdaf50268aac881dcd3ad52f7448323233)