Remove V82 compatibility layer used in printing.
There are very weird things going on with the diff API, I suspect it should
not be written the way it is, but rely on more algebraic inputs enforcing
more invariants. I do not have time to fix this right now, so for the
time being this trivial patch should be enough.