Remove chunks of compatibility layers from ssrfwd.
This one patch is tricky because we have to play explicitly with the goal
states, otherwise it breaks the semantics of ssr intropatterns. Indeed,
the Proofview.Goal.of_tactic implicitly drops the state attached to a goal
because there is no way to do otherwise.

I suspect that this is merely uncovering a bug elsewhere, but for the sake
of advancing in the removal of compatibility layers I am just emulating the
previous of_tactic behaviour.