Correctly remove defined evars from the future goal state.
There is no point in keeping defined evars inside that state, since its only
purpose is to track the set of undefined evars to turn into goals. Given
that the FutureGoals asymptotics is pretty bad this would lead to quadratic
behaviours when repeatedly creating and instantiating evars.

This does not solve the underlying quadratic behaviour due to FutureGoals,
but it makes it much harder to hit. Now one needs to effectively have many
future goals lying around for it to happen.

Fixes #13798: Quadratic slowdown in performance of implicit arg holes.