Simplify the sorting of eauto subgoals.
The current comparison algorithm is lexicographic with depth, priority and
number of generated subgoals. Due to the way we compute the depth and priority,
assumptions and introductions were always guaranteed to come first.

Therefore this patch should not change the observable semantics, assuming that
the sorting algorithm was stable on assumption tactics.