Provide an explicit primitive to raise non-zero Ltac failures.
The rationale is that non-zero level failing is an Ltac-specific feature.
We have a few tacticals around that try to catch this level and reraise it
with a lower level, but this hardly makes sense for ML tactics. Only Ltac
programs should ever observe the level, and even there this feature is often
used to work around deficiencies of the expressivity of Ltac.

As a matter of fact, most ML callers of the tclFAIL primitive call it with a
level set to 0, which clearly indicates an API issue.

This PR removes the tclFAIL level argument, and provides a tclFAILn function
to pass an arbitrary number when needed.