Do not unconditionally record backtraces in CoqIDE.
This was introduced by f3f165b and is probably a remnant of debugging. We
should not record backtraces in general, this can be expensive. The other
removed line was broken in general since it was printing the backtrace
after an arbitrary function call.

(cherry picked from commit ef6a2a59f12eb0df430c30fd6049b92d5942acf8)