Remove internal goal-printing API from Printer.
There are two kinds of goal printing functions: debug printers and UI display.
The Printer module was exposing low-level functions used internally to implement
coqtop UI display. It is asking for trouble to give this to the user, since
it should never be used for debugging. The only legitimate users of this API
are UIs, and coqtop only calls the toplevel funtion. Better remove them.