Rely on the subgoal approximation to compute eauto hint sorting.
The goal of this change is to be able to statically order eauto hints without
having to execute them. This is slightly backwards incompatible but the differences
in behaviour are only observable for Extern hints. Since we cannot decide a
priori the number of goals generated by an Extern hint we rely on the priority
instead.