Fixes #9555: substitution of module type expression wrongly done over parameters.
When defining a module type F with params X:T by an expression M, the
self-references to M in M must be replaced by F (i.e. implicitly "F X")
but no substitution should be done in the parameters.

We fix it by moving the substitution on the body of the functor.

By inlining the previous call to subst_modtype_and_resolver (which
calls subst_body), one sees that it calls subst_signature on the unary
substitution "M |-> F" with empty delta-resolver.

The codomain should however be substituted, so, we use a variant of
subst_signature which sustitutes also the codomain of the
delta-resolver of M.

(cherry picked from commit 7e3d11946f0214d88b9bd92322011c322f397004)