Fixes #9579 and fixes support for inlining level in module parameters.
Issue #9579 was about a (locally) wrong representation of the
algebraic form of functors and functorial types in declaremods.ml.

Inlining levels in module parameters were otherwise treated
inconsistently:
- In "Module F (X:T [at inline level 10]) := ...", the level was
  discarded (before calling Mod_typing.translate_module) except to check
  possible <: constraints (in Declaremods.check_subtypes).
- In "Module F (X:T [at inline level 10]). ... End F.", the level was
  taken into account (via Safe_typing.add_module_parameter).

Note that the inline level (or "!", or "no inline") has an effect on
the immediate module type it is attached to. The effect is visible
only if the type is an applied functorial type and, if so, it tells
how to inline the argument of the application.