Little optimization in computing the composition of module substitution.
What we actually need is composition of closures, so no need for
keeping useless bindings.

That is, in a closure M[subst], the result of M[subst][subst'] is
simply M[subst[subst']].

The tricky point is that we need to build explicitly an identity
substitution to enforce an initial closure.