Use anonymous primitive projections in the VM.
This patch changes the PROJ VM opcode to only take the projection index as
argument, rather than both the index and the high-level Projection.Repr.t
representation. This allows removing the code for global registering of
projections in the VM, thus reducing the size of relocations. Since this
also makes the bytecode slightly shorter, this is a net gain in memory and
vo size.

The only place where this has an effect is during reification of accumulated
projections. One could think that this makes the process more expensive since
we now have to reconstruct the projection, but in reality we were already
accessing the environment there to fetch the inductive block (and doing so
several times, actually). The new code only accesses the environment once,
so even there this should be faster on average.