Remove the flashbar display for found strings in CoqIDE.
This feature was both dubious and very broken. Most of the time it was not
displaying anything due to the info bar being overwritten by other irrelevant
data. Worse, when the file contained non-ASCII characters the index table
contained garbage and, in addition to not displaying anything, output
an annoying "Anomaly: Occurrence of Find string not in table" on stderr.
To add insult to injury, this was also very slow on big files because it would
copy the text in its entirety to perform the search.

(cherry picked from commit 7fdb21026688092ad0ea2c4766289e70200bf6d7)