Fixes #15403: inlining not propagated along modpath aliases.
The invariants of the code for delta-resolver composition are complex
and I don't understand them well. At least, there was a discrepancy
when composing either a delta-resolver [Top.NAT'=>Top.E] or
delta-resolver [Top.NAT'.t=>Top.E.t] with a substitution
{Top.E |-> Top.NAT [Top.NAT.t=>inline(Some _)]}: the second case was
returning [Top.NAT'.t=>inline(Some _)] but not the first case. The
discrepancy was in gen_subst_delta_resolver, which
subst_codom_delta_resolver was calling and we fix the first case.