Using strongly wellfounded fixpoints in tests with loops in erasable subterms,
Checking guard in previously ignored erasable or inert subterms of fixpoints.

We do that by splitting guard errors into categories:
- fatal recursive calls to non guarded arguments
- recursive calls on descendants of variables bound in subterms of a
  redex and thus liable to be instantiated by reduction of the redex +
  not applied enough recursive calls which also may become enough
  applied after reduction.

We also maintain a stack of redexes so that when a variable is bound
in the subterm of a redex, we directky know which redex needs to be
reduced to trigger the instantiation of the variable.

In situations of the form "C(if c then x else y)" where x is a
recursion variable, y needs the unfolding of a redex C[], we use a
rich notion of subterm which remembers both that we need to reduce the
redex to check a possible recursive call on y, but that we need to
check that we don't also make a recursive call on x even w/o reducing
the redex. So, the union of specifications of subterms of a match/if
is now a set.

For the beta and zeta-redex cases, we choose to check the argument for
possibly fatal non guarded calls and reduce anyway. Another strategy
could have been to reduce only when a subterm of the redex needs a
reduction (but the stack call is then a bit complicated to manage).

Note: We could have been more tolerant in subterms with recursive
calls on internal uninstantiatable variables and on not applied enough
recursive calls which cannot be applied more (e.g.
"Fixpoint foo (n : nat) := forall x : foo = foo, True"). To accept
these calls, "check_erasable_inert_subterm_rec_call" should be
renamed into "check_inert_subterm_rec_call" to emphasize that it
becomes more general and checks in types (Ind, Prod, types in
Fix/Cofix/Lambda/Array) should call it in place of "check_rec_call".