[findlib] compat with dune (legacy plugin names)
We let the developer select the legacy loading mechanism by writing
  Declare ML Module "legacy_plugin_name:package-name.plugin"
The legacy method is available while building the library itself, so
that dune < 2.9.2 can work. Still the plugin is installed as a
regular findlib library and clients can load it as such. In particular
the DECLARE PLUGIN directive uses the new name, not the legacy one.

Plugins in the coq-core package can be loaded by just using their
legacy name, since public name is known (and appended on the fly).