Make our only use of the OCaml Ephemeron API compatible with OCaml 5.0.
Instead of using the low-level Ephemeron API, we simply use a dummy weak
hash table where the key structure is trivial. This might be marginally
more costly but it is unlikely to make any difference in practice.

We also seize the opportunity to put the arguments of is_ground_env in the
standard order, i.e. the environment before the evarmap.