Slightly tweak the API of Tactics.general_elim_clause.
Despite taking a full clausenv, this function was already expecting the
evarmap of the corresponding clausenv to have been set as the global one
from the tactic monad. As such, there is no point in taking a clausenv,
we simply take its signature and retrieve the set of metas from the current
evarmap.