Fixing bug in computing the default name of an index in recursion schemes.
An environment was incorrect, resulting in not using the type to
generate the name of a variable in this type.

I don't know how to strictly be compatible with the previous behavior
though (see test PrintingParentheses.v). For instance, forcing
"Inductive eq A (a : A) : forall y : A, Prop" would also forces some
names to be "y" where it could have been generated from the type
instantiating A.