Use a faster irrelevance check for constants and constructors in conversion.
We store a map of irrelevant constants and inductive types in the environment,
with the hope that on average these maps will be small compared to the global
environment. This makes checking irrelevance O(1) when it is not used, so this
is effectively a zero-cost abstraction. This seems to remove most of the
overhead of the previous patches in the wild, except for unimath.