Function find_inductive finds types s.t. mind_finite=Finite (i.e. Inductive).
This simplifies the description of the condition about when a fixpoint
is allowed (see #407, 7ceed3c76e), and let the tactic "fix" follows
the same rule as the term constructor "fix".

Currently, mind_finite=Finite means given with the keyword Inductive
but it could eventually become truly-recursive Inductive (excluding
e.g. prod).