Ltac2 now compiled with "-noinit"
We now compile all of Ltac2 with the "-noinit" flag. The only file
that is left out is Ltac2.Notations, due to a dependency on `f_equal`
from the `cc_plugin` and `easy` from Coq.Init.Tactics.

Ltac2.Ltac2 is now accompanied by Ltac2.Noinit which the former
`Require`s. This allows projects which use "-noinit" to import
Ltac2.Noinit and use Ltac2. For other projects, Ltac2.Ltac2 should
have the same behaviour as before.