[micromega] Optimise the algorithm to perform interval analysis
The goal is to find an interval for a monomial x1^a1....xj^aj
using existing bounds c1 + c2.x1^b1....xj^bj >= 0

The old algoritm was blindly multiplying bounds until reaching the
     maximal multi-degree of the initial monomials.
It was therefore generating an exponential number of spurious bounds.

The new algoritm works as follows.
To bound a monomial m, it solves a system of equations in the ai of the form
(m1)^a1.....(mj)^aj = m
Given a solution (a1,...,an), it takes the bound of each mi to the power ai

Finding all the solution of the equations may also be costly.
The advantage is that we only generate bounds for monomials of interest.