Deprecate the Typeclasses Filtered Unification flag.
This flag is hopelessly broken. It enables a fast syntactic check for
the application of TC hints. This is not per se a bad idea, but the
implementation is completely crazy, since it considers evars in the goal
as rigid for the check. That is, as soon as the the TC search must find
a proof whose goal mentions an evar, this goes boom immediately.

As such, I suspect that nobody is using this flag in the wild. I believe that
we should reimplement the feature from scratch for efficiency reasons, but in
the meantime it does not hurt to signal to our end-users that they should not
rely on the current implementation.