Use a faster normalization for proof extraction in rwprocess_rule.
Instead of calling simpl, which performs a step of nf-βι, we use full-blown
whd-all. This might cost a bit more because of δ-unfoldings, but it will
compensate by not normalizing whole parts of the term that will never
reach the head. On terms that already reduce to a pair it should not matter.