Line timings summarised at end of bench
We change render_results to be a regular OCaml executable instead of a
toplevel script.

We put the pretty printer for tables into its own OCaml library. We
also modify it so that alignments can be specified for all columns.

We create an executable render_line_results which displays the top
speed ups and slow downs.

This then gets called in the bench.sh script.