Remove the univ_of_sort function from the API.
This is a trap convincing the user that universes and sorts are isomorphic
and can be safely translated between one another. But the reality is that this
should not be the case and the user should match on the inductive description
of sorts instead. Soon universe will be restricted to algebraic expressions,
making this function impossible to write.