Do not rely on Prop existing as a root for the Print Sorted Universes command.
Up to now, this command was printing Prop < Set despite never having been true,
except for implementation details. The net effect of this patch is to remove
this line from the ouptut. This has the consequence that the script checking
the number of universes in the test-suite is now returning one less universe.