Correctly handle functor arguments for module subtyping check.
The invariant of the subtyping function is that the intended subtype is
expected to be part of the current environment. Unfortunately, it was
added too late for the check of the argument types to be performed correctly.

I do not have a reproducible example but this is a variant of #8937 I have
discovered while tweaking the conversion algorithm used for subtyping. This
test only passes in master because it takes the fast path for conversion based
on equality, but the terms conversion is provided with are referring to
unbound inductive types in the current environment.