Make the kernel fully universe-inference-free.
The only code that was relying on inferring constraints was the one for
modules. We correctly compute the constraints in Declaremods and simply check
them in Safe_typing.

This is duplicating a bit of tricky code between the kernel and the upper layers,
as well as redoing some of the work. We were already performing some checks in
the upper layers just to throw them aways, so the added redundance is not so
big. Furthermore we should be able to reduce the added code by checking the
constraints more aggressively.