Fast complete check for universe variance.
If a level does not appear syntactically in a term, this term does not contribute
to the cumulativity relation for this level. This assumes that constant bodies
cannot refer to local universes, but this is enforced by the cumulativity check.
This patch introduces an underapproximation to skip the variance computation
when in this situation.