Unconditionally load opaque proofs in coqchk even in norec mode.
This was an optimization, but it turns out that opaque proofs from dependencies
are still accessed by functor instantiation and module inclusion. Therefore,
as a simple fix we just load them as well. We could have tried to be more clever
by either loading it on the fly or carrying around additional data for delayed
typing, but it is probably not worth it both in terms of complexity and
efficiency.