Allow for overriding of .csdp.cache file for micromega
We check for a new env var MICROMEGA_CACHE_FILE which if set to
"cache_file" will read the file "cache_file.cache", copy it to a new
file "cache_file.cache.new" and continue using the new file as the
persistent csdp cache, as .csdp.cache was being used before.

If the env var is not set, the old behaviour of reading and writing to
.csdp.cache is kept.

This will allow us to manage the micromega cache in a controlled way
which will be needed for porting the test-suite to dune.