[proof engine] Forbid tactic code to raise exceptions
Motivated by the recent discussion on Zulip about exceptions and
tactics, we move forward with the porting of tactic code to the
monadic proof engine and forbid tactics to raise non-critical
exceptions.

Instead, tactics calling exceptional code should catch and reify the
exception locally and use the proper proof engine API `tclZero`.

I think this is a very long due cleanup.

I hope @ppedrot likes this!

As of now, this is a draft, and not yet ready for benchmark. Stdlib
compiles, test suite is close to pass.

This should be combined with the work on PR #XXXX that adds exception
static analysis to Coq's codebase.

Next step is to remove `wrap_exceptions`, but that will require some
more tweaks as it appears is some critical paths.