Slight tweak of heuristic for template poly inference.
We check that an inductive without an explicit Prop arity has been lowered to
Prop after the fact. In this case, we make it template. This contrasts with
inductive where the return arity has been explicitly set to Prop, in which
case it is not template.

Normally this should be essentially equivalent to the previous code, but
this is quite tricky.