[mltop] Improve and fix the legacy plugin loading method.
The current implementation of the "legacy" loading for plugins doesn't
store the plugin name that was used in the `Summary` when saving a
`.vio` / `.vo` file. This is necessary in some cases such as when
doing `vio2vo` without installing Coq.

We simplify the logic here and store the full information in the
summary of how the plugin was loaded, which allow us to remove the
current hack we have, and makes the code path a bit more
deterministic.

The old loading method was "relocatable", in the sense -I controlled
the paths for the legacy plugin loading too for now.

This should improve the compat with build systems (such as Dune) that
don't support yet findlib plugin building.

We also do some general refactoring of the code, in particular:

- we make the digests mandatory
- we simplify the native / byte plugin selection logic

There are some other improvements we can make in future PRs:

- there are a few non-deterministic plans such as "if file X is here,
  do this" which is always suspicious, we could improve that: in
  incremental builds either a file is a dep

- also, `coqc` and `coqdep` have different semantics for `-I`, I'd be
  great if we could improve this.