[build] Consolidation of build system infrastructure
We replace both the dune-based `(coq.theory ...)` and the "legacy"
make-based build systems for .vo into a single setup, based on our own
Dune rule generation using a coq_dune tool.

This requires a bootstrap process to generate the rule files, until
Dune gets support for `(include stdlib_rules.sexp)` (hopefully soon)

By disabling the `(coq.theory ...)` stanza on `theories/dune` we lose
compositional building of .vo libs (for OCaml we are still fine, which
our main use case), this is not really used as of today so we are
fine.

This should supersede #15545 , avoid many troubles that #15220 had
to endure, and close #15482 .

Legacy makefiles will be removed in the next commit, closing a few
more bugs hopefully.

At some point some of this code should be moved to a common loadpath
library to compute native target names, and maybe some other stuff.

Some other changes in `Nativelib`:

- We fix incorrect early initialization of `Boot.Env.t`

- We extend the path passed to the ocamlc compiler as to search for
  files in the dir where the .vo is.

  That can be reverted if we pass all the explicit `-nI` paths.

  We could just get rid of `.coq-native`, additionally.