Stop inferring template universes in the kernel.
Instead we ask the upper levels to provide them and we only check that they
are indeed template.

The new code is also slightly more lenient since it is more consistent in what
is allowed to be template w.r.t. cumulative inductive types. Now, a template
universe need not appear in the return arity anymore, it is enough for it
to appear as a sort parameter. Similarly, Prop-valued inductive types can
now have template parameters.