[libobject] Store definition location in the .vo files
This has been a long-requested feature that we need to implement it
properly for users; after a bit of discussion with Ali we decided to
go ahead.

For now, the patch is a proof of concept, however it does seem to work
fine, but in general we should make the location of objects a generic
property, not an ad-hoc one as done here. In particular, locations
need to be handled better (and likely first-class) in libobject as to
cover some corner cases.

I'm willing to shepherd anyone willing to work on it.

At least, the `CInfo` refactoring did pay off pretty well, so this was
not too hard.

As a POC of the feature, I've enabled location printing in `Search`
and `Locate`:

```
Locate add.

Constant Coq.Init.Nat.add
  File "./Nat.v", line 47, characters 9-12
  (shorter name to refer to it in current context is Nat.add)
```