[vernac] Fix nested proof handling in interpretation.
It seems that #14779 could have introduced a bug here, the proof is
always dropped so the stack has constant size of one.

Unfortunately we cannot test this well as the STM overrides the vernac
implementation for nested proofs.

This will make coq-lsp and the simple compiler happy on nested proofs tho.