[coqtop] Don't initialize the boot environment on -boot
This was a subtle bug, as `coqtop` was actually respecting `-boot` on
the main path but trying to find `coqlib` for printing the version in
the header.

We should have a better way to locate the `revision` file, (it
shouldn't require searching for `Init/Prelude.vo`) but for now we just
patch the bug.

Needed for #13364, fixes #16276 .