[pcoq] Don't use physical equality to determine Pcoq.freeze worklist
I've been observing huge parsing slowdowns in interactive use for a
while, and I finally found out that these are due to physical equality
in `share_tails` missing a lot of same-entries, thus creating a huge
work list for setting the parsing state in `Pcoq.freeze`.

This experimental commit uses a more precise equality, which is a 50x
win over a file interactive reparse, but indeed, it seems that we
should go for a more radical solution, which to really put the parsing
state in a reference we can just switch in O(1).