[libobject] Common metadata for declared objects
This is a first implementation of CEP https://github.com/coq/ceps/pull/65

We add a common metadata object to atomic libobjects, so we can store
documentation, location, names, and other interesting attributes that
are needed by clients to implement functionality.

As we allow storing names of the objects, we can index them by name to
provide efficient reverse lookup.

We export an experimental API for clients to consume this metadata.

PR #16261 is implemented on top of this one.