Fixes #16303: evar defined before its instance is checked.
Was revealed by commit 26a5b65df in PR #14797.

In the same move, we introduce check_and_define_evar_instance which
does both the unification of types of evar and body and the
instantiation.

We consistently do the instantiation before type-checking (as it was
done most of the cases before this commit). Doing it the other way
round breaks a few examples.