Issue #16262: do not override generic format if not reserved and only printing.
This is an experimental change of heuristic about when "Notation" is
liable to modify a generic printing rule.

The argument for the change is that if we want to change the generic
format of an only-printing rule, we can still do it with "Reserved
Notation", while before this commit, "Notation" forces a change of
generic rule (i.e. a rule associated to the parsing/printing structure
of a notation) even if the intent is only to declare a specific
printing rule (i.e. a rule associated to a given interpretation of a
notation).

Co-authored-by: Gaëtan Gilbert <gaetan.gilbert@skyskimmer.net>