Be stricter about substitution of evar instances.
Forecasting the change of representation of evar instances, we now assume
we cannot access the default parts of the instance without access to an
evarmap. This means that all functions from the kernel now raise an anomaly
when trying to mess with potential Var nodes of evar instances, and that their
counterpart in the upper layer now take an additional evarmap argument.