Simplify error message for "cannot move _ it occurs in the type of _"
This is sometimes incorrect as the issue can be the body of the variable.

We just say "it occurs in _" instead of specifying whether it's in the
type or the body because the code is too complicated to get that
information easily (it is not clearly available in moverec_toright).

Fix #16364