Do not use unsafe casts for hypothesis computation in Declare.
There is a cryptic comment near that piece of code about the requirement to
be unsafe not to count the variables appearing in the type of evars, but it
does not make sense in the current code. Indeed, for evar-normalized terms
the set of variables computed by the unsafe functions and the safe ones must
coincide. Given that 1. we have an evarmap around and 2. it is a code smell
to manipulate terms with free evars without an evarmap, this patch is a net
gain.