Do not reason up to aliasing in Evarsolve.assoc_up_to_alias.
This particular quirk was introduced by bfefc3f to allegedly solve an issue
with the aliasing mechanism that was long forgotten to the mist of time.
Now, given the current implementation I am not even sure how to trigger this
piece of code in a way that makes an observable difference. Given that this
prevents an efficient implementation of the lookup, I think that it is
reasonable to simply remove it.