Partially revert #9139.
This change was a footgun because it was inciting the caller to use the
provided environment and evarmap to perform printing of the thunk, although
that data could have been unrelated to the current one. Given that the trace
mechanism is supposed to be called from within the monad, always, it is
natural to simply fetch the relevant information from the ambient pure state
rather than fiddle with broken contraptions.