Tweak nf_evar so that it does not expand default instances.
This seems to be critical for efficiency, because we unconditionally
call nf_evar on sight in many places in pretyping amongst others.
Expanding the instances makes every subsequent call to EConstr.kind
and friends slower.

In particular, evar expansion is needed in ssreflect to correctly handle evar
surgery. This change actually makes mathcomp about 4% *faster* than master,
when it was 8% *slower* before.