[loadpath] Small API cleanup
We remove the "library location" information bit as it is only used
for printing and introduces a dependency on `Library` which is
unnecessary, as in my opinion we'd like loadpath resolution to become
an independent library (so it can be shared with other tools).

This bit was added more than 20 years ago!