Fix printing of `rewrite` when SsrSyntax is imported
Ssreflect has some incompatible syntax with normal Coq, see
https://coq.inria.fr/refman/proof-engine/ssreflect-proof-language.html#compatibility-issues.
In particular `rewrite a, b` no longer works and has to be replaced with
`rewrite -> a, -> b`. During printing, this is also a problem.

In this PR, we check wether `SsrSyntax` is loaded, an if so, we print `rewrite`
with forward `->` arrows.

Testcase:
```
Ltac t a b := rewrite a, b.
Print t.
(* Used to print rewrite a, b and still does *)
Require Import ssreflect.
Print t.
(* Used to print rewrite a, b and now prints rewrite -> a, -> b *)
```