Non rectypes Nativevalues.t using GADT
This uses morally a single `Obj.magic` to produce a witness `(t, t -> t)
eq` for abstract `Nativevalues.t` (in practice 4 `Obj.magic` for varying
arities for efficiency reasons, based on looking at generated assembly).

Rectypes is still needed for the generated files as `let Refl = (t_eq
: (t, t -> t) eq = ...` only works when rectypes is on.

However if we also deal with the misc uses in coq (constr, logic
monad, etc) we should be able to remove -rectypes from the dune flags
and only use it when calling ocamlopt for native compile.