Move the Prod case from native atoms to values.
We had to cheat a bit to abuse the value representation of blocks of tag 0
in the native compiler. There used to be two possible kinds of 0-tagged blocks:
- Either they had size >= 2 in which case they were accumulators coded as closures,
where the tag 0 was the well-known hack to make them behave as an ADT constructor
at runtime.
- Or they had size 1, in which case they were persistent arrays. The fact this
works was abusing the internal representation of persistent arrays, which are
a reference on some ADT with 2 non-constant constructors.

In order to extend this, we simply wrap Prod values into a dummy 0-tagged block
and encode them as a 3-uple with a tag that is distinct from the ones defined
by the PArray implementation.