Remove the Vuniv_level constructor of VM value kinds.
This was not making any sense, because the code was confusing the untyped
representation of the value and the view of this datum as an inductive
classification. If anything, this would probably have led to segfaults if
used improperly, but thankfully the few callers were doing trivial things
with the cast functions.