Tweak the ElimArity error to account for the change of case representation.
The printer was never using the provided judgement in case the arity error
was None, so we move it there. We also remove the arity_error field since
it was computed systematically out of the other fields.

This allows using this error for low-level issues with the case typing rule,
when the arity is ill-formed.