[coqdep] Be more deterministic w.r.t. the plugin loading mode
If the legacy mode is set, we don't even attempt to use findlib to
resolve anything, we also won't emit a META dependency, even if this
dependency is in scope.

This is morally the right thing to do, should reduce non-determinism,
and helps a bit to isolate the legacy code more.

This should help with problems such as
https://github.com/ocaml/dune/issues/5833