Hooks have no default value
Hooks are meant to be a workaround to lack of cross-file recursive
module, and as such a hook is meant to refer to 1 specific function.

Hooks with default values were either meant to be set to arbitrary
functions by plugins in which case they should be refs, or the default
was not actually used.