Refine the evar-evar fast path criterion in Evarsolve.
When unifying ?x[a] ~ ?y[b], the algorithm tries to define y := x[φ(a)], where
φ(a) is some clever inversion of the arguments of x so that this is correct. The
previous fast path was simply checking that a and b were trivial, in which case
it would pick φ(a) := a, plus some additional checks for candidates. Actually,
the only thing we care about a is that it makes sense in the default context
of y. Thus, we generalize the fast path only check for a to be closed and not
mentioning y. The previous behaviour is a special case, and as far as I can tell
the new one is compatible with the non-fast-path branch.