Use a kernel environment instead of a special-purpose structure in VM compilation.
Since the introduction of a O(log n) rel lookup in kernel environments, the original
purpose of the Renv type was probably made moot. Furthermore, some parts of the
code were relying on the environment for reduction, but the one passed to the function
was the global one, leading to potential anomalies (and even maybe unsoundness).